simplify iframe interface
