+++ src/screens/Apps/Escrow/Index.tsx
@@ -1,4 +1,4 @@
-import { useContext, useEffect, useMemo, useState } from 'react'
+import { useContext, useEffect, useState } from 'react'
 import { bytesToHex, hexToBytes } from '@noble/hashes/utils.js'
 import { hashes } from '@noble/secp256k1'
 import { sha256 } from '@noble/hashes/sha2.js'
@@ -6,11 +6,13 @@ import { Transaction } from '@scure/btc-signer'
 
 import Header from '../../../components/Header'
 import { NavigationContext, Pages } from '../../../providers/navigation'
-import { ArkadeIdentityHandlers, ArkadeIframeHost } from './ArkadeIframeHost'
+import { ArkadeIframeHost } from './ArkadeIframeHost'
 import { WalletContext } from '../../../providers/wallet'
 import { FlowContext } from '../../../providers/flow'
 import { Network } from '@arkade-os/boltz-swap'
 import { AspContext } from '../../../providers/asp'
+import { base64 } from '@scure/base'
+import makeMessageHandler from './RpcHandler'
 
 // Needed to sign the message, perhaps to be lifted up int App structure?
 hashes.sha256 = sha256
@@ -50,20 +52,26 @@ export default function AppEscrow() {
     return svcWallet?.getBalance()
   }
 
-  async function sign(tx: Transaction, inputIndexes?: number[]): Promise<Transaction> {
+  async function signArkTransaction(base64Tx: string, base64Checkpoints: string[]) {
     if (!svcWallet) throw new Error('Wallet not initialized')
-    return svcWallet.identity.sign(tx, inputIndexes)
+    const tx = Transaction.fromPSBT(base64.decode(base64Tx), { allowUnknown: true })
+    const checkpoints = base64Checkpoints.map((_) => base64.decode(_))
+    const signedTx = await svcWallet.identity.sign(tx)
+    const signedCheckpoints = await Promise.all(
+      checkpoints.map(async (cp) => {
+        const signed = await svcWallet.identity.sign(Transaction.fromPSBT(cp, { allowUnknown: true }), [0])
+        return base64.encode(signed.toPSBT())
+      }),
+    )
+    return {
+      signedTx: base64.encode(signedTx.toPSBT()),
+      signedCheckpoints,
+    }
   }
-  async function signerSession() {
-    throw new Error('Not implemented')
-  }
-  async function signin(hexToSign: string) {
+  async function signLoginChallenge(hexToSign: string) {
     if (!svcWallet) throw new Error('Wallet not initialized')
     const signatureBytes = await svcWallet.identity.signMessage(hexToBytes(hexToSign), 'schnorr')
     return bytesToHex(signatureBytes)
-  }
-  async function signout() {
-    return { ok: true }
   }
   async function fundAddress(address: string, amount: number): Promise<void> {
     setSendInfo({
@@ -78,19 +86,14 @@ export default function AppEscrow() {
     navigate(Pages.SendDetails)
   }
 
-  const handlers: ArkadeIdentityHandlers = useMemo(
-    () => ({
-      sign,
-      signerSession,
-      signin,
-      signout,
-      getXOnlyPublicKey,
-      getArkWalletAddress,
-      getArkWalletBalance,
-      fundAddress,
-    }),
-    [svcWallet, navigate, setSendInfo],
-  )
+  const handlers = makeMessageHandler({
+    getXOnlyPublicKey,
+    signLoginChallenge,
+    getArkWalletAddress,
+    getArkWalletBalance,
+    signArkTransaction,
+    fundAddress,
+  })
 
   const allowedOrigins = Object.values(BASE_URLS)
     .filter((_) => _ !== null)
@@ -100,7 +103,7 @@ export default function AppEscrow() {
     <>
       <Header text='Escrow on Ark' back={() => navigate(Pages.Apps)} />
       {escrowAppUrl !== null && (
-        <ArkadeIframeHost src={escrowAppUrl} allowedChildOrigins={allowedOrigins} handlers={handlers} />
+        <ArkadeIframeHost src={escrowAppUrl} allowedChildOrigins={allowedOrigins} messageHandler={handlers} />
       )}
     </>
   )
